Move bd::settings from runner.h into settings.h
